requirements.txt
amqp==2.6.1
billiard==3.6.4.0
celery==4.4.7
click==7.1.2
configparser==4.0.2
contextlib2==0.6.0.post1
Flask==1.1.4
flask-peewee==3.0.4
Flask-WTF==0.14.3
gunicorn==19.10.0
importlib-metadata==2.1.3
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
pathlib2==2.3.7.post1
peewee==3.15.4
python-dotenv==0.18.0
pytz==2022.6
redis==3.5.3
scandir==1.10.0
six==1.16.0
typing==3.10.0.0
validate-email==1.3
vine==1.3.0
waitress==1.4.4
Werkzeug==1.0.1
wtf-peewee==3.0.4
WTForms==2.3.3
zipp==1.2.0
